chore: add type annotations
